Add filter data by hash url
If we pass some filtering paramater in the url, the dashboard applies that filter when the data is loaded.
We assume that the argument name will be data and in the url we set the returned value.
For example: filter=data.landmark.indexOf("Eiffel")<0 should filter all the data whos landmark got the
string "Eiffel".